update: improve accessibility
